- Added sql/database -> datagridview functionality (aka the ui will now load from sql data hard coded files) - Added method to convert animal object to a list of strings (UNTESTED)
Note: Since the datagridview now uses the sql database instead of hardcoded files, you will need to setup mysql. There is a link to the installer in the github readme for this project. All that is required is that you install it, the rest (db setup, config, making tables, etc...) is automagically taken care of in the code.